+++ src/moby/image.go
@@ -1,4 +1,4 @@
-package main
+package moby
 
 import (
 	"archive/tar"
