+++ src/moby/schema.go
@@ -1,4 +1,4 @@
-package main
+package moby
 
 var schema = string(`
 {
