+++ src/moby/trust.go
@@ -1,4 +1,4 @@
-package main
+package moby
 
 import (
 	"crypto/tls"
