+++ src/moby/trust_test.go
@@ -1,4 +1,4 @@
-package main
+package moby
 
 import "testing"
