+++ src/moby/util.go
@@ -1,0 +1,6 @@
+package moby
+
+var (
+	// MobyDir is the location of the cache directory which should be set by the caller
+	MobyDir string
+)
