+++ src/moby/util_unix.go
@@ -1,0 +1,11 @@
+// +build !windows
+
+package moby
+
+import (
+	"os"
+)
+
+func homeDir() string {
+	return os.Getenv("HOME")
+}
