+++ src/moby/util_windows.go
@@ -1,0 +1,9 @@
+package moby
+
+import (
+	"os"
+)
+
+func homeDir() string {
+	return os.Getenv("USERPROFILE")
+}
